tools/axiom/requirements.txt
PyYAML>=6.0
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
streamlit-plotly-events>=0.0.6
reportlab>=4.0.0
kaleido>=0.2.1
jinja2>=3.1.0
jsonschema>=4.0.0